UIEH-1005 Rename files
